requirements.txt
Django
psycopg2-binary
djangorestframework
djangorestframework-gis
django-cors-headers

